requirements.txt
fastapi>=0.136.0
httpx>=0.28.1
langfuse>=4.5.0
openai>=2.32.0
python-dotenv>=1.2.2
uvicorn>=0.45.0
